improve camera performance
